move getEnsureOIDCProviderInstalledTask: tasks -> tasks/platforms
